Organize root-level components into layout/ subdirectory
Move 13 sidebar layout components from the components root into
components/layout/ for better organization. Delete empty legacy
layouts/ directory.

Closes #119
