+++ resources/js/components/layout/AppContent.vue
