+++ resources/js/components/layout/AppSidebar.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
-import NavFooter from '@/NavFooter.vue';
-import NavMain from '@/NavMain.vue';
-import NavUser from '@/NavUser.vue';
+import NavFooter from '@/layout/NavFooter.vue';
+import NavMain from '@/layout/NavMain.vue';
+import NavUser from '@/layout/NavUser.vue';
 import { Sidebar, SidebarContent, SidebarFooter, SidebarHeader, SidebarMenu, SidebarMenuButton, SidebarMenuItem } from '@/ui/sidebar';
 import { type NavItem } from '@/types';
 import { Link } from '@inertiajs/vue3';
