+++ resources/js/components/layout/PlaceholderPattern.vue
